Mostly remove unused comments
