+++ whisper_demos/whisper_demos/stream.py
@@ -41,7 +41,7 @@ class TranscriptSubscriber(Node):
         super().__init__('transcript_subscriber')
 
         # Control what elements of the transcript are displayed
-        self.declare_parameter('threshold', 2.0)
+        self.declare_parameter('threshold', 1.0)
         
         # Subscribe to the audio transcript topic
         self.subscription = self.create_subscription(
